tailcalling vm
Base implementation to test for correctness (all 'testes/' pass). Some
functions are explicitly outlined to validate generated assembly and
parts of this design will likely be redone.

vmbreak may also be modified in the future to allow inline-assembly
(e.g., labels) to be played with in a much easier fashion.